move webapp template from example to package
