+++ .deta/prog_info
@@ -1,1 +1,1 @@
-{"id":"a93803fc-50ca-4082-898c-d4d34ee96936","space":41455,"runtime":"python3.9","name":"example-bot","path":"uubd8u","project":"a0f8c57y","account":"144798365827","region":"eu-central-1","deps":["aiogram==3.0.0b6","aiogram-deta==1.0.0a0","deta[async]==1.1.0a2","fastapi"],"envs":["TELEGRAM_TOKEN","TELEGRAM_SECRET","DETA_RUNTIME","DETA_PROJECT_KEY"],"public":false,"log_level":"off","cron":""}
+{"id":"2d571e18-263d-4fae-9f55-738175f04ebe","space":41537,"runtime":"python3.9","name":"AiogramDeta","path":"f8pa4y","project":"a0gws2rb","account":"144798365827","region":"eu-central-1","deps":["aiogram==3.0.0b6","aiogram-deta==1.0.0a0","deta[async]==1.1.0a2","fastapi"],"envs":["TELEGRAM_TOKEN","TELEGRAM_SECRET"],"public":true,"log_level":"debug","cron":""}
