+++ .deta/state
@@ -1,1 +1,1 @@
-{"LICENSE":"3b87bea5a6816fd89422c694cf3a1b3cdeb0bae7c35412cf8effc447646b9136","README.md":"4f3122da8fc9fc9947711d492321146ed997bc0ac754c960f2d3a594d10234d8","aiogram_deta/__about__.py":"b188991173b898532b7ec8a6f444cbefc3f6f71cce8cd7ac474001ebd19fa581","aiogram_deta/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","aiogram_deta/storage.py":"6b68c1bc7fd6322c0ace440fe6e9e622ef6548c90c6e8022679ad70b02846d81","deta.json":"386cebac079220444fd5ea8d27d1d81073ecba9e9a4d4f15a2aad7d51b32b6cc","dist/aiogram_deta-1.0.0a0-py3-none-any.whl":"59ca72a8fca40cd5cd043beda44a9f199429d28f428408323cc15ffcefbfede0","dist/aiogram_deta-1.0.0a0.tar.gz":"ea16fbba2fd4213de78a228887b9261a469d9c1dfa1ab63ef8e1a769136f87ac","example_bot/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/factory.py":"102afe35f20d7785b70226c03c9c35cfc72fabb4a7a17941b72750cef5f21bb0","example_bot/bot/handlers/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/handlers/echo.py":"021f8dbe4faaf256c32b8cc03e2721655cac49bede5593759a5464a8df01e25c","example_bot/bot/handlers/form.py":"82705e0c724e0d2b479c9531d57222a0afb21856a25870fc355801e7e6207855","example_bot/web/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/web/factory.py":"187541343fe687d6eb809c76d81331104b99c26509a9ebb88b9c615ef96c4509","example_bot/web/routes/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/web/routes/webhook.py":"9f81c3ea2f4cd65a3004d84037ea2a3266ee2dc36300808ab3d525f9c16ec23c","example_bot/web/stubs.py":"5e8239d2326babd2971a13b99c600b7c9a5b184684996f50d0c98b6093e0a551","main.py":"f03a7b4270c129fb26734a4c925dee6f303259daf749c5878e80f01ac39fbcde","pyproject.toml":"4ad7260ffde94ab880987ecb6b4968e3f8541f627426efb66b65af07156ad74f","requirements.txt":"76837ef5dabf5e3c364a323153ae5f8a6e35ad19c99804440f9f5a4fb2d315c0"}
+{"LICENSE":"3b87bea5a6816fd89422c694cf3a1b3cdeb0bae7c35412cf8effc447646b9136","README.md":"fe0a609b0017703c12d78cf6e5b4229e13d24ef58e6a1e6bfe99d16abfaff6c6","aiogram_deta/__about__.py":"b188991173b898532b7ec8a6f444cbefc3f6f71cce8cd7ac474001ebd19fa581","aiogram_deta/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","aiogram_deta/storage.py":"6b68c1bc7fd6322c0ace440fe6e9e622ef6548c90c6e8022679ad70b02846d81","aiogram_deta/web/__init__.py":"ebe3c21ceb5b099ac26c0c4ea1e1afa40b711a6b0da256c00e8c9fa833ab37cd","aiogram_deta/web/factory.py":"39a714570daa6accb85e89d208ee7e4e95f963b3064325c9afba643c4d47d1a7","aiogram_deta/web/stubs.py":"5e8239d2326babd2971a13b99c600b7c9a5b184684996f50d0c98b6093e0a551","aiogram_deta/web/webhook.py":"83762a838a02edb6352cec59e0c2f548e9d0a02904651253f55cfa09a3a46fd8","deta.json":"386cebac079220444fd5ea8d27d1d81073ecba9e9a4d4f15a2aad7d51b32b6cc","example_bot/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/factory.py":"102afe35f20d7785b70226c03c9c35cfc72fabb4a7a17941b72750cef5f21bb0","example_bot/bot/handlers/__init__.py":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","example_bot/bot/handlers/echo.py":"021f8dbe4faaf256c32b8cc03e2721655cac49bede5593759a5464a8df01e25c","example_bot/bot/handlers/form.py":"93108544f6d10eaf6fa87e6dcb5c8165bf73201b20d50db5d2885bcefa1496ed","main.py":"f10b8a072a6c25216c01a83608735a0762dbd418e340e94bdba9f59a91ef0194","pyproject.toml":"02310241e6625c723a3290b58ee65c68529db216de2c93ace5113545708001f8","requirements.txt":"76837ef5dabf5e3c364a323153ae5f8a6e35ad19c99804440f9f5a4fb2d315c0"}
