+++ .gitignore
@@ -4,3 +4,4 @@ dist/
 .pytest_cache
 .venv
 .env
+.deta
