+++ aiogram_deta/web/__init__.py
@@ -1,0 +1,3 @@
+from aiogram_deta.web.factory import create_app
+
+__all__ = ["create_app"]
