+++ aiogram_deta/web/stubs.py
