+++ main.py
@@ -3,7 +3,7 @@ from os import getenv
 from deta import Deta
 
 from example_bot.bot.factory import create_bot, create_dispatcher
-from example_bot.web.factory import create_app
+from aiogram_deta.web import create_app
 
 deta = Deta()
 bot = create_bot(token=getenv("TELEGRAM_TOKEN"))
